Fix compilation issues when compiling with clang
clang had an issue with deducing template parameters for the
apply_function parameter of func_apply in the OMEs. Instead the
initialisation now wraps the function pointer explicitly in the correct
std::function type.
Moreover, the tests for func_plusfunc_omx used constexpr std::log, which
is a C++26 feature and also tripped up clang. This commit removes the
constexpr keyword in this test case.
Thanks for Gavin Salam for reporting these issues.